+++ crates/bcr-wallet-core/src/lib.rs
@@ -16,6 +16,7 @@ use cashu::{CurrencyUnit, KeySetInfo, MintInfo, MintUrl};
 use cdk::wallet::{MintConnector as MintCon, types::TransactionId};
 use chrono::Utc;
 use error::{Error, Result};
+use secp256k1::{Keypair, SECP256K1};
 use std::{
     collections::{HashMap, HashSet},
     str::FromStr,
@@ -130,9 +131,10 @@ impl AppState {
                 return Err(Error::InvalidNetwork(w_cfg.network, self.cfg.network));
             }
 
-            if w_cfg.mnemonic != self.cfg.mnemonic {
+            let keypair = keypair_from_mnemonic(&self.cfg.mnemonic);
+            if w_cfg.pub_key != keypair.public_key() {
                 tracing::error!(
-                    "Mnemonic mismatch: wallet {wid} has a different mnemonic than the one given via config"
+                    "Key mismatch: wallet {wid} has a different pubkey than the one given via the config mnemonic"
                 );
                 return Err(Error::InvalidMnemonic);
             }
@@ -153,7 +155,7 @@ impl AppState {
                 w_cfg.name,
                 w_cfg.network,
                 w_cfg.mint,
-                w_cfg.mnemonic,
+                self.cfg.mnemonic.clone(),
                 LocalDB::Keep,
                 Self::DB_VERSION,
                 self.cfg.same_mint_safe_mode,
@@ -690,7 +692,8 @@ async fn build_wallet(
     same_mint_safe_mode: SameMintSafeMode,
     db: Arc<redb::Database>,
 ) -> Result<ProductionWallet> {
-    let seed = mnemonic.to_seed("");
+    let seed = seed_from_mnemonic(&mnemonic);
+    let keypair = keypair_from_mnemonic(&mnemonic);
     // retrieving mint details
     let client = ProductionConnector::new(mint_url.clone());
     let keyset_infos = client.get_mint_keysets().await?.keysets;
@@ -745,11 +748,22 @@ async fn build_wallet(
         (debit_pocket, credit_pocket),
         name,
         wallet_id,
-        mnemonic,
+        keypair.public_key(),
         beta_clients,
         Box::new(|url| Box::new(crate::mint::HttpClientExt::new(url))),
         same_mint_safe_mode,
     )
     .await?;
     Ok(new_wallet)
+}
+
+// converts a mnemonic to a secp256k1 keypair
+fn keypair_from_mnemonic(mnemonic: &bip39::Mnemonic) -> Keypair {
+    let seed = seed_from_mnemonic(mnemonic);
+    let (key, _) = seed.split_at(secp256k1::constants::SECRET_KEY_SIZE);
+    Keypair::from_seckey_slice(SECP256K1, key).expect("key to be correct size")
+}
+
+fn seed_from_mnemonic(mnemonic: &bip39::Mnemonic) -> [u8; 64] {
+    mnemonic.to_seed("")
 }